nlp_processing/requirements.txt
spacy
transformers
torch
fastapi
uvicorn
python-dotenv
